Add Booking logic

+++ CarRental.Common/Classes/Motorcycle.cs
@@ -7,12 +7,12 @@ public class Motorcycle : IVehicle
     public string RegNo { get; init; }
     public string Make { get; init; }
     public int Odometer { get; init; }
-    public decimal CostPerKilometer { get; init; }
+    public double CostPerKilometer { get; init; }
     public VehicleTypes VehicleType { get; init; }
-    public decimal CostPerDay { get; init; }
+    public double CostPerDay { get; init; }
     public VehicleStatuses Status { get; init; }
 
-    public Motorcycle(string regNo, string make, int odometer, decimal costPerKilometer, VehicleTypes vehicleType, decimal costPerDay, VehicleStatuses status)
+    public Motorcycle(string regNo, string make, int odometer, double costPerKilometer, VehicleTypes vehicleType, double costPerDay, VehicleStatuses status)
     {
         RegNo = regNo;
         Make = make;
